remove heuristic around directory listing
If rclone detect a directory filter it will prefer listRwalk instead of listR.

The idea is to skip listing the content of unnecessary directory.

In practice for our S3/GCS with millions of objects it makes a huge consumption in memory as rclone will
build the tree in memory.

See discussion here:
https://forum.rclone.org/t/high-memory-usage-of-s3-during-listing/35376/10